Combined SCFG testing files
