+++ numba_rvsdg/tests/test_scfg.py
@@ -4,6 +4,8 @@ from textwrap import dedent
 from numba_rvsdg.core.datastructures.scfg import SCFG
 
 from numba_rvsdg.tests.test_utils import SCFGComparator
+from numba_rvsdg.core.datastructures.basic_block import BasicBlock
+from numba_rvsdg.core.datastructures.labels import ControlLabel
 
 
 class TestSCFGConversion(SCFGComparator):
@@ -98,6 +100,25 @@ class TestSCFGConversion(SCFGComparator):
         for case in cases:
             scfg = SCFG.from_dict(case)
             self.assertEqual(case, scfg.to_dict())
+
+
+class TestSCFGIterator(SCFGComparator):
+
+    def test_scfg_iter(self):
+        expected = [
+            (ControlLabel("0"), BasicBlock(label=ControlLabel("0"),
+                                           _jump_targets=(ControlLabel("1"),))),
+            (ControlLabel("1"), BasicBlock(label=ControlLabel("1"))),
+        ]
+        scfg = SCFG.from_yaml("""
+        "0":
+            jt: ["1"]
+        "1":
+            jt: []
+        """)
+        received = list(scfg)
+        self.assertEqual(expected, received)
+
 
 if __name__ == "__main__":
     main()
